requirements-gpu.txt
tornado>=6.1
setuptools
pydantic
wheel
google-auth-oauthlib
httpx==0.23.0
fairseq==0.12.2
tensorboardX
faiss_cpu==1.7.3
ffmpeg_python==0.2.0
ffmpy==0.3.1
websockets>=10.0
gradio==3.34.0
librosa==0.9.1
elevenlabs
gTTS==2.3.2
wget
matplotlib==3.7.2
mega.py==1.0.8
gdown
noisereduce==2.0.1
unidecode
numba==0.57.1
numpy==1.23.5
onnxruntime
onnxruntime_gpu==1.15.1
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
pandas==2.0.3
praat-parselmouth==0.4.2
PySimpleGUI==4.60.5
pyworld==0.3.3
requests==2.31.0
resampy==0.4.2
scikit_learn==1.3.0
scipy==1.11.1
yt_dlp==2023.7.6
pyngrok==4.1.12
sounddevice==0.4.6
soundfile==0.12.1
tb_nightly==2.14.0a20230803
tensorboard==2.13.0
torchcrepe
torch_directml==0.2.0.dev230426
torchgen>=0.0.1
tqdm==4.65.0
